pyproject.toml
[tool.poetry]
name = "nicholascooks"
version = "0.1.0"
description = ""
authors = ["Nicholas Zolton <nicholaszolton@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.115.4"}
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
structlog = "^24.4.0"
pyjwt = "^2.9.0"
alembic = "^1.14.0"
psycopg = {extras = ["binary", "pool"], version = "^3.2.3"}
pytest = "^8.3.3"
cryptography = "^43.0.3"
# mangum = "^0.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
